+++ packages/process/lib/Polysemy/Process/Data/PtyError.hs
@@ -1,0 +1,5 @@
+module Polysemy.Process.Data.PtyError where
+
+data PtyError =
+  CreationFailed
+  deriving stock (Eq, Show)